% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatch_expansion.R
\name{bycatch_expansion}
\alias{bycatch_expansion}
\title{bycatch_expansion is the primary function for fitting bycatch models to time series of takes and effort}
\usage{
bycatch_expansion(time = NULL, events = NULL, cov = NULL, effort = NULL,
  coverage = NULL, family = c("poisson"))
}
\arguments{
\item{time}{Numeric, e.g. years}

\item{events}{Integer vector of takes (events)}

\item{cov}{Optional matrix of covariates}

\item{effort}{Metric of fishing effort to be used in estimation of mean bycatch rate}

\item{coverage}{Observer coverage (0 - 100) used for binomial expansion}

\item{family}{Observation error distribution, defaults to Poisson}
}
\value{
list of the data used to fit the model, the expanded bycatch generated via the fit and simulations, and the fitted stan model
}
\description{
bycatch_expansion is the primary function for fitting bycatch models to time series of takes and effort
}
