% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stream_summary.R
\name{get_stream_summary}
\alias{get_stream_summary}
\title{get_stream_summary provides a summary of monitoring coverage and bycatch by stream}
\usage{
get_stream_summary(fitted_model, alpha = 0.05)
}
\arguments{
\item{fitted_model}{Data and fitted model returned from fit_bycatch() with multi-stream monitoring}

\item{alpha}{The alpha level for the credible interval, defaults to 0.05}
}
\value{
data frame with summary statistics for each monitoring stream
}
\description{
get_stream_summary provides a summary of monitoring coverage and bycatch by stream
}
\examples{
\donttest{
d <- data.frame(
  Year = 2002:2014,
  Takes_obs = c(0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 0, 0, 0),
  Sets_obs = c(200, 180, 150, 350, 250, 270, 180, 150, 400, 350, 280, 180, 250),
  Takes_em = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0),
  Sets_em = c(150, 120, 140, 250, 180, 190, 120, 100, 300, 270, 200, 130, 190),
  Sets_total = c(1000, 950, 900, 1800, 1200, 1300, 900, 750, 2000, 1800, 1400, 950, 1300)
)

fit <- fit_bycatch(Takes_obs ~ 1,
  data = d, time = "Year",
  effort = "Sets_obs",
  takes_em = "Takes_em",
  effort_em = "Sets_em",
  effort_total = "Sets_total",
  family = "poisson"
)

get_stream_summary(fit)
}
}
