% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total.R
\name{get_total}
\alias{get_total}
\title{get_total is a helper function to return a matrix of total estimated bycatch}
\usage{
get_total(fitted_model)
}
\arguments{
\item{fitted_model}{Data and fitted model returned from estimation}
}
\value{
matrix (MCMC draws x time steps) of posterior predictive values for total bycatch (observed + unobserved)
}
\description{
get_total is a helper function to return a matrix of total estimated bycatch
}
\examples{
\donttest{
d <- data.frame(
  "Year" = 2002:2014,
  "Takes" = c(0, 0, 0, 0, 0, 0, 0, 0, 1, 3, 0, 0, 0),
  "expansionRate" = c(24, 22, 14, 32, 28, 25, 30, 7, 26, 21, 22, 23, 27),
  "Sets" = c(391, 340, 330, 660, 470, 500, 330, 287, 756, 673, 532, 351, 486)
)
fit <- fit_bycatch(Takes ~ 1,
  data = d, time = "Year",
  effort = "Sets",
  family = "poisson",
  expansion_rate = "expansionRate",
  time_varying = FALSE
)
total <- get_total(fit)

# Calculate total bycatch summaries
total_mean <- colMeans(total)
total_quantiles <- apply(total, 2, quantile, probs = c(0.025, 0.975))
}
}
