% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred_distribution.R
\name{plot_pred_distribution}
\alias{plot_pred_distribution}
\title{plot_pred_distribution plots the distribution of bycatch prediction for a specific year}
\usage{
plot_pred_distribution(
  fitted_model,
  year,
  xlab = "Predicted Bycatch",
  ylab = "Frequency",
  breaks = NULL
)
}
\arguments{
\item{fitted_model}{Data and fitted model returned from fit_bycatch().}

\item{year}{Specified year for the prediction distribution}

\item{xlab}{X-axis label for plot}

\item{ylab}{Y-axis label for plot}

\item{breaks}{Breaks for bars in the plot, defaults to 1}
}
\value{
A ggplot object.
}
\description{
plot_pred_distribution plots the distribution of bycatch prediction for a specific year
}
\examples{
\donttest{
d <- data.frame(
  "Year" = 2002:2014,
  "Takes" = c(0, 0, 0, 0, 0, 0, 0, 0, 1, 3, 0, 0, 0),
  "expansionRate" = c(24, 22, 14, 32, 28, 25, 30, 7, 26, 21, 22, 23, 27),
  "Sets" = c(391, 340, 330, 660, 470, 500, 330, 287, 756, 673, 532, 351, 486)
)
fit <- fit_bycatch(Takes ~ 1,
  data = d, time = "Year",
  effort = "Sets",
  family = "poisson",
  expansion_rate = "expansionRate",
  time_varying = FALSE
)
plot_pred_distribution(fit, year = 2002)
}

}
