\name{bycatch-package}
\alias{bycatch-package}
\alias{bycatch}
\docType{package}
\title{
\packageTitle{bycatch}
}
\description{
\packageDescription{bycatch}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bycatch}
\packageIndices{bycatch}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{bycatch}

Maintainer: \packageMaintainer{bycatch}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
#simple examples of the most important functions
}
