% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatch-package.R
\docType{package}
\name{bycatch-package}
\alias{bycatch-package}
\alias{bycatch}
\title{The 'bycatch' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
