% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{expand does the expansion from a fitted object}
\usage{
expand(fit, coverage = NULL, control = list(sigfig_multiplier = 100,
  mcmc_samples = 1000, maxX = 20000))
}
\arguments{
\item{fit}{the estimated bycatch rate from a call to fit}

\item{coverage}{the coverage rate (0-100). This can be a scalar if all values are the same, or a vector corresponding to each row of the data frame used in fitting}

\item{control}{list of 3 elements used for control: sigfig_multiplier (used for adjusting precision of estimates, defaults to 1000), mcmc_samples (number of mcmc samples randomly selected from the posterior for expansion), maxX (upper bound for drawing from latent bycatch events, also related to sigfig_multiplier)}
}
\value{
expanded_estimates the estimates of expanded estimates across rows of the dataframe and mcmc samples
}
\description{
expand does the expansion from a fitted object
}
\examples{
\donttest{
d = data.frame("Year"= 2002:2014,
"Takes" = c(0, 0, 0, 0, 0, 0, 0, 0, 1, 3, 0, 0, 0),
"expansionRate" = c(24, 22, 14, 32, 28, 25, 30,  7, 26, 21, 22, 23, 27),
"Sets" = c(391, 340, 330, 660, 470, 500, 330, 287, 756, 673, 532, 351, 486))
fit = fit_bycatch(Takes ~ 1, data=d,
time="Year", effort="Sets", family="poisson", time_varying = FALSE)
expanded = expand(fit, coverage = d$expansionRate)
}
}
