% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{extract_log_lik_for_loo}
\alias{extract_log_lik_for_loo}
\title{Extract log-likelihood matrix for loo, removing NA observations}
\usage{
extract_log_lik_for_loo(fit)
}
\arguments{
\item{fit}{A bycatch model fit object}
}
\value{
A matrix suitable for loo::loo()
}
\description{
Extract log-likelihood matrix for loo, removing NA observations
}
