% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expanded.R
\name{plot_expanded}
\alias{plot_expanded}
\title{plot_expanded is makes plots of the expanded bycatch estimates, accounting for observer coverage and effort}
\usage{
plot_expanded(fitted_model, xlab = "Time", ylab = "Events")
}
\arguments{
\item{fitted_model}{Data and fitted model returned from bycatch_expansion()}

\item{xlab}{X-axis label for plot}

\item{ylab}{Y-axis label for plot}
}
\value{
plot called from ggplot
}
\description{
plot_expanded is makes plots of the expanded bycatch estimates, accounting for observer coverage and effort
}
