% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expanded.R
\name{plot_expanded}
\alias{plot_expanded}
\title{plot_expanded is makes plots of the expanded bycatch estimates, accounting for observer coverage and effort}
\usage{
plot_expanded(
  fitted_model,
  expanded_estimates,
  xlab = "Time",
  ylab = "Events",
  include_points = FALSE
)
}
\arguments{
\item{fitted_model}{Data and fitted model returned from estimation}

\item{expanded_estimates}{values returned from a call to expansion()}

\item{xlab}{X-axis label for plot}

\item{ylab}{Y-axis label for plot}

\item{include_points}{whether or not to include raw bycatch events on plots, defaults to FALSE}
}
\value{
plot called from ggplot
}
\description{
plot_expanded is makes plots of the expanded bycatch estimates, accounting for observer coverage and effort
}
\examples{
\donttest{
d = data.frame("Year"= 2002:2014,
"Takes" = c(0, 0, 0, 0, 0, 0, 0, 0, 1, 3, 0, 0, 0),
"expansionRate" = c(24, 22, 14, 32, 28, 25, 30,  7, 26, 21, 22, 23, 27),
"Sets" = c(391, 340, 330, 660, 470, 500, 330, 287, 756, 673, 532, 351, 486))
fit = fit_bycatch(Takes ~ 1, data=d, time="Year",
effort="Sets",
family="poisson",
time_varying = FALSE)
expanded = expand(fit, coverage = d$expansionRate)
plot_expanded(fitted_model=fit,
expanded_estimates = expanded,
xlab="Year",
ylab = "Fleet-level bycatch",
include_points = TRUE)
}
}
