% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted.R
\name{plot_fitted}
\alias{plot_fitted}
\title{plot_fitted makes plots bycatch estimates (lambda of Poisson), accounting for effort but not accounting for observer coverage}
\usage{
plot_fitted(fitted_model, xlab = "Time", ylab = "Events")
}
\arguments{
\item{fitted_model}{Data and fitted model returned from bycatch_expansion()}

\item{xlab}{X-axis label for plot}

\item{ylab}{Y-axis label for plot}
}
\value{
plot called from ggplot
}
\description{
plot_fitted makes plots bycatch estimates (lambda of Poisson), accounting for effort but not accounting for observer coverage.
}
